refactor: command 패키지 이동
