+++ src/main/java/chess/command/End.java
@@ -1,4 +1,4 @@
-package chess.view.command;
+package chess.command;
 
 public class End implements Command {
