+++ src/test/java/chess/command/EndTest.java
@@ -1,9 +1,9 @@
-package chess.view;
+package chess.command;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import chess.view.command.Command;
-import chess.view.command.End;
+import chess.command.Command;
+import chess.command.End;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
